up::json: Added builder class up::json::builder and removed overloaded constructor for up::json::value with std::initializer_list<value>.
On one hand, it should be simple to construct JSON objects with
initializer lists. On the other hand, an initializer_list for
up::json::value is ambigious, because it might also be a
up::json::object in some cases.

For this reason, there is no a builder class to construct
up::json::array and up::json::object with initializer_list.