NEW: up::error for lean exceptions.
up::exception depends on up::fabric that depends on up::string that
depends on throwing exceptions. To be able to break the cycle,
up::error introduces lean exceptions (aka errors). The latter should
be used for up::string, overflow checks, etc, and the up::exception
should be changed to inherit from up::error.