up::throwable: Refactor root exception once again.
The major difference between up::throwable and up::exception should be
that the latter supports introspection with up::insights. The new
design for up::throwable (formerly known as up::error) supports both
parents and fields, and the same design should be used for
up::exception. In contrast to up::exception, the fields in
up::throwable can only be accessed with debugging tools and not from
the program itself.